Minor restructuring from the parse file
Move some functions to other locations where it makes more sense
